+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
     - Home screen added.
+    - Screen2 added, but wrapping text around image does not work :(
 
 
 ## [0.1.1] - 2020-03-22